requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
streamlit==1.44.1
